implemented MapBlocksToUnitsTest

+++ src/main/java/ipb/pt/timetableapi/solver/SizeConstant.java
@@ -4,5 +4,5 @@ public class SizeConstant {
     public static final double SIZE_5 = 5.0;
     public static final double SIZE_2_5 = 2.5;
     public static final double SIZE_0_5 = 0.5;
-    public static final int UNIT_DURATION = 30;
+    public static final long UNIT_DURATION = 30L;
 }
